requirements.txt
numpy==1.24.2
pandas==1.5.3
db-dtypes==1.0.5
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sendgrid==6.9.7
google-cloud-bigquery==3.6.0
google-cloud-storage==2.7.0
flask==2.2.3
